# fix: add support for symlinked files and dirs