@RazrFalcon 2024-07-30
This change is very expensive. System fonts loading is up to 50% slower on macOS. I cannot accept it.

We have to ignore duplicated fonts.

Also, I don't know how Nix works, but can we follow symlinks only for files, ignoring directories? Or will it not work?