@jcdickinson 2024-07-30
@RazrFalcon Nix uses symlinks very heavily, so it needs full support. I'll think about how to optimize this more.