@jcdickinson 2024-07-30
@RazrFalcon, I have made the following changes:

* Always support adding symlinked files. This codepath is avoided if no symlinks are found, so should be zero cost.
* Allow directory symlink searching to be disabled with the `dir-symlink` feature flag.
* So long as a symlink is never encountered, the `dir-symlink` should be very close to no different.
* More syscalls are unavoidably involved when encountering a symlinked file or directory, so these will always be more expensive.
* Add a criterion benchmark.
  
Here are the results from my machine, your machine will likely/hopefully show almost no difference.

## Before

*To test: make the second `fn canonicalize` always return the first value.

```sh
cargo run --example find-system-font --release
Loaded 464 font faces in 9ms (19μs per).
Error: 'Times New Roman' not found.
```

```sh
cargo bench
load system fonts       time:   [15.849 µs 15.897 µs 15.947 µs]
```

## Without dir-symlink

```sh
cargo run --example find-system-font --release
Loaded 464 font faces in 10ms (23μs per).
```

```sh
cargo bench
load system fonts       time:   [19.768 µs 19.841 µs 19.916 µs]
                        change: [+24.201% +24.807% +25.420%] (p = 0.00 < 0.05)
```

## With dir-symlink

```sh
cargo run --example find-system-font --release
Loaded 504 font faces in 13ms (27μs per).
```

```sh
cargo bench
load system fonts       time:   [22.803 µs 22.822 µs 22.842 µs]
                        change: [+14.579% +15.025% +15.461%] (p = 0.00 < 0.05)
```