@jcdickinson 2024-07-31
I just realized that canonicalization is not required when dir symlinks aren't being followed, changes pushed:

## vs previous commit

```sh
load system fonts       time:   [25.457 µs 25.495 µs 25.536 µs]
                        change: [-20.722% -20.594% -20.454%] (p = 0.00 < 0.05)
```

## vs no changes at all

```sh
load system fonts       time:   [25.978 µs 25.988 µs 25.998 µs]
                        change: [+8.4845% +8.7590% +9.0271%] (p = 0.00 < 0.05)
```