@jcdickinson 2024-07-31
Fixed a bug which basically turned the file-only implementation into a noop. Results on my system:

## No changes

```
load system fonts       time:   [24.501 µs 24.517 µs 24.533 µs]
...
Loaded 81 font faces in 2ms (28μs per).
```

## With file symlink support

(and symlinked files). It's *faster?* I wonder if it's behaving favorably with some kernel cache.

```sh
load system fonts       time:   [21.658 µs 21.673 µs 21.689 µs]
                        change: [-11.678% -11.599% -11.512%] (p = 0.00 < 0.05)
...
Loaded 122 font faces in 2ms (24μs per)
```

## With dir support

```
load system fonts       time:   [25.220 µs 25.236 µs 25.252 µs]
                        change: [+3.1805% +3.4172% +3.6518%] (p = 0.00 < 0.05)
...
Loaded 122 font faces in 3ms (27μs per).
```