@drupol 2024-07-31
Hey, Nix and Typst user here!

I'm looking for that fix too since the build environment is made of symlinks, using `fontdb` 0.18 on the development branch of Typst is simply not working any more. I made a temporary project to show the issue if you're really interested at https://github.com/drupol/typst-font-issue

Looking forward for the next `fontdb` release, including this fix!