@jcdickinson 2024-07-31
Pushed changes to address PR comments. Before:

```
Loaded 81 font faces in 2ms (28μs per).
```

After:

```
Loaded 122 font faces in 3ms (27μs per).
```

I'm very confused as to why there is no performance difference, even though there are more syscalls and hashing the PathBuf (which I did encounter last night, maybe a :hankey: implementation?). Still, no complaints.