@RazrFalcon 2024-08-01
>using fontdb 0.18 on the development branch of Typst is simply not working any more

Symlinks were never supported.