@RazrFalcon 2024-08-01
Only if Typst was handling them on its own before.