@EpicEricEE 2024-08-01
It was working before because prior to #65, [`Path::is_file`](https://doc.rust-lang.org/std/path/struct.Path.html#method.is_file) and [`Path::is_dir`](https://doc.rust-lang.org/std/path/struct.Path.html#method.is_dir) were used, which traversed symlinks automatically. It was then changed to [`DirEntry::file_type`](https://doc.rust-lang.org/std/fs/struct.DirEntry.html#method.file_type), which doesn't do that anymore.