@jcdickinson 2024-08-06
Made some changes to get it to compile on MSRV: https://github.com/jcdickinson/fontdb/actions/runs/10259900356