@RazrFalcon 2024-08-06
Looks good now. I don't see any significant performance differences on macOS loading 1396 font faces.